requirements.txt
requests==2.32.4
pillow==11.3.0
openpyxl==3.1.5
python-docx==1.2.0
PyPDF2==3.0.1